Remove unnecessary uint8 cast in EIP-712 encoding
The cast to uint8 is redundant since the EIP-712 type string already
specifies uint8 transferType, making the encoding identical without
the explicit cast.

+++ contracts/libraries/PeripherySigningLib.sol
@@ -122,7 +122,7 @@ library PeripherySigningLib {
                     hashBaseDepositData(swapAndDepositData.depositData),
                     swapAndDepositData.swapToken,
                     swapAndDepositData.exchange,
-                    uint8(swapAndDepositData.transferType),
+                    swapAndDepositData.transferType,
                     swapAndDepositData.swapTokenAmount,
                     swapAndDepositData.minExpectedInputTokenAmount,
                     keccak256(swapAndDepositData.routerCalldata),
